image signer, verifier and loader scripts

+++ img-verify.sh
@@ -7,9 +7,15 @@ sigf=$(mktemp)
 
 openssl x509 -in $1 -pubkey -noout > $pubk 
 
-echo $3 | base64 -d --wrap=0 > $sigf 
+if [ -e $3 ] 
+then 
+    base64 -d --wrap=0 $3 >> $sigf
+else 
+    echo $3 | base64 -d --wrap=0 > $sigf 
+fi
 
-openssl dgst -sha256 -verify $pubk -signature $sigf $2
+result=$(openssl dgst -sha256 -verify $pubk -signature $sigf $2)
 
 rm -rf $pubk $sigf
 
+echo $result
